requirements_dev.txt
black==23.3.0
duckdb
flake8
GitPython
google-cloud-bigquery>=3.6.0,<4.0.0
google-cloud-storage>=2.7.0,<3.0.0
pyarrow
pytablewriter
retry
sh>=2.0.0
